pyproject.toml
[project]
name = "tgn-whisperer"
version = "0.1.0"
description = "Podcast transcription and processing for TGN, WCL, and Hodinkee Radio"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "xmltodict",
    "requests",
    "invoke",
    "anthropic",
    "python-dotenv",
    "loguru",
    "tenacity",
    "httpx[http2]",
    "zensical",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "tldextract>=5.1",
    "PyYAML>=6.0",
    "tqdm>=4.66",
    "rich>=13.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["app"]
